整数型の微調整

+++ my/q04_01.cpp
@@ -1,6 +1,6 @@
 #include "gtest/gtest.h"
 
-int tribo(const int N) {
+long long tribo(const int N) {
     if (N < 0) {
         throw std::invalid_argument("argument must not be negative");
     }
